requirements.txt
openai
tweepy
pillow
python-dotenv
atproto